tests: merge test_fddp.py and test_solver.py
